Added BGM and SFX to the game
sound effects and explain:
hurt: Plays when the player gets hurt(health-1).
explosion: Plays when the bomb explodes.
ignite-bomb: Plays when the bomb is placed.
enemy-die: Plays when the enemy dies.
enemy-hurt: Plays when the enemy gets hurt.
victory: Plays when the player finishes one level.
steel-pipe: Plays when the player dies.
button-click: Plays when you click on a button.